[VFX] CustomHLSL Extend Arbitrary Buffer Support to Texture
➕ Extension from this PR https://github.cds.internal.unity3d.com/unity/unity/pull/45766 and relative to this JIRA [VFXG-477](https://jira.unity3d.com/browse/VFXG-477) linked to [POI-1353](https://jira.unity3d.com/browse/POI-1353) for 6.1

This PR extends support of read/write data to texture, the great advantage of this kind of buffer is being able to allocate them in asset folder through RenderTexture and/or Custom Render Texture.

Behind the scene, it uses the same pattern and storage than RWBuffer. The VFXExpressionUsage can take as input either a GraphicsBuffer or any texture.

🎁 This PR is also allowing parameter of type `Texture2D` directly (where we were forced to always use the macro `VFXSampler2D`)