[VFX] Fix subgraph block nesting
Fix for https://jira.unity3d.com/browse/UUM-83521

Nesting subgraph blocks was not possible anymore and all blocks in subgraph block contexts appeared as "invalid" (detoured in red)
After investigation it appeared that this had never really working. However before "the regression" the issue was less obvious: Subgraph blocks could accept too many blocks regardless of their compatibility, and the red detour feedback was not working correctly. There was a big ambiguity in code about context type and compatible context types. I tried to refactor this a bit so that it's clearer and less hacky.

- Accepting blocks should work with all contexts now. For subgraph block context, it can only accept a block that is compatible with *all* its suitable contexts. 
- Invalid visual feedback should work and be refreshed correctly now (so for instance when switching the suitable context setting of the subblock context)