Fix Depth attachment texture for DX11
Fixed the CameraDepthAttachment Texture for Dx11, which was rendering as black earlier. 

The issue is the Camera Depth Attachment texture is referring to the depth buffer and when the shader is trying to access the texture the texture is black.
This mainly happens because of 2 reasons:
1. When we try to set the active depth buffer texture, the code unbinds the texture(passes null value to PSSetShaderResources)
2. As the depth buffer is active, we need to set OMSetRenderTargets for depth as null to be able to access it. 

Bug: https://jira.unity3d.com/browse/UUM-64316
Port: https://jira.unity3d.com/browse/UUM-85190