[URP] Fixing an issue where Transparent Receive Shadows toggle wasn't working (UUM-79471)
Fixes UUM-79471.
The issue is that for Main Light Shadows, it changes the shadow fade to accomplish the goal of making transparent objects not receive shadows. While Additional Light Shadows do this by setting the shadow strength to 0 and marking every light as a non shadow casting. This discrepancy causes issue for users making custom shaders that don't use the Shadow Fade for main light shadows. To fix the issue, shadow strength is now set to 0 for the transparent passes if the toggle is disabled (so transparent objects should not receive shadows).

This PR also cleans up the shadow codebase a little bit, similar to #56069