[Port] [6000.0] RenderGraph.AddBlitPass doesn't work correctly with texture array slices
Fix UUM-77828
The render graph compiler was not correctly handling sub-resources as attachments both in c# and c++. This could cause rendering to slice x and y to be merged into a single render pass that rendered only to slice x. This is fixed by comparing not only the resource but also the array slice and mip index.

Also fixes the following minor issues discovered during investigation of this bug:
- Add Slice/Mip data to the render graph debugger which made it difficult to debug this
- Add some extra validation to the descriptor to avoid silently failing when creating invalid texture descriptors
- Renderpass C++ api validation was broken and considered some valid setups to be invalid since it was not comparing the aray slice and mip when testing attachments for equality.