[Port] [6000.0] [HDRP] Better material creation for PBR Sky Custom Material
Currently, the only way to use the default shader graph in the HDRP package for custom PBR Sky is to use the menu (that no one knows about). 
![image](https://media.github.cds.internal.unity3d.com/user/1764/files/2beff507-dad9-40c0-bf69-468041483d5a)

And the default material is in the HDRP package making it readonly.
Currently : 
![image](https://media.github.cds.internal.unity3d.com/user/1764/files/35eb2961-ba4b-4331-be82-c43cc185c112)

After the PR: 
What this PR does is add the pattern that we already have elsewhere (custom pass, water surface, water decal) to basically create a NEW material based on a shader graph that is in the HDRP package using the NEW button.
![image](https://media.github.cds.internal.unity3d.com/user/1764/files/11ad1dab-19c2-4459-9618-f777b25a93b7)


After clicking on new, a new folder is created using the name of the current scene, assign it and highlight it. 
This is the material created, it should help users not knowing how to use the PBR Sky master stack by default by not starting from an empty master stack. 

![image](https://media.github.cds.internal.unity3d.com/user/1764/files/b79f107d-7b51-42c0-a838-5396d1f9e71a)

This PR also fix a minor discrepancy, between setting no material and overidding the material property and setting no material and not overriding the property. The two gave 2 different render whereas they should have been the same. (cf [commit](https://github.cds.internal.unity3d.com/unity/unity/pull/56435/commits/df0fc3e234358fdbf5210fcaf00b369baada6da1))