[Port] [6000.0] [HDRP] Empty template starts with incorrect Physically Based Sky (incorrect exposure in later frames)
This PR fixes the issue of removing the celestial data twice resulting in a black sky cubemap.

The cleanup calling order was changed after this PR(https://github.cds.internal.unity3d.com/unity/unity/pull/49443) was applied.
The ordering change caused the removal of the celestial data for sky rendering in the middle of the render pipeline.

So it has been fixed to call the cleanup only once to maintain the celestial body data.