Multiview Per View Viewports Part 2
This PR includes fixes for corner cases previously identified in Phase 1 for the Multiview Per View Viewports (MVPVV) extension for Quest. The goal of Part 2 is to get MVPVV working with user-defined viewports and addressing corner cases. Again, MVPVV is only applied to the last render pass in each frame, when the eye textures are being rendered to.

MVPVV allows developers to set multiple viewports and scissors per viewport. In the context of VR rendering, this enables the drivers to skip shader invocations (and rendering work) for screen areas outside of the user’s view (these are the nasal regions that the user can't see within the headset). Note that it only works with symmetric Field of View (FOV).

We removed the "-enable-vulkan-mvpvv" argument and instead now rely on the XRBootConfig/MVPVV checkbox in the xr.sdk.oculus settings (only available in Unity 6.1 and above) to determine whether to execute the MVPVV paths in the engine.

This change is also needed in xr.sdk.oculus to pass the two different XR scissors per eye and adds the MVPVV checkbox to Oculus Plugin Settings: https://github.cds.internal.unity3d.com/unity/xr.sdk.oculus/pull/644. These engine-side changes contained within this PR do not take effect without the plugin-side changes.