Fix URP Lit shader for mobile WebGL
Fix URP Lit shader for WebGL on Adreno GPUs.
Some mobile browsers have a small GL_FRAGMENT_MAX_UNIFORM_VECTORS. The Galaxy Tab S8 has a limit of 256, while a desktop browser has 1024. This causes URP Lit shaders to fail to load on these devices, resulting in the object not rendering.

Shadows.hlsl normally puts shadow values into a CBuffer, but it does not do this for GLES3, due to some devices having performance issues with large uniform buffers. We need to do this for WebGL to avoid the uniform vector limit, so the shader was updated to include WebGL for using a CBuffer for the shadow values.