[6000.0] Fix picking and outline for all BatchRendererGroup draw types
Backport of https://github.cds.internal.unity3d.com/unity/unity/pull/56758

Previously, the internal method `OnDrawCommand()` used for BatchRendererGroup picking and outline expected a pointer to a `BatchDrawCommand` (i.e. a regular direct draw), which it expects to be able to modify, at least for the material and mesh fields.

When non-direct (procedural and/or indirect) draws were added to BRG, this code was not changed, and draws were simply skipped if they were not of the regular direct type.

The existing `OnDrawCommand` implementations did not actually require the draw to be direct. Instead, they need to:
* Select the Material the draw is performed with, which can differ from the original Material.
* Skip over some draws, e.g. in case the draw is present in an ignore list.
* Know the submesh index so it can be communicated to the shader.

This PR modifies `OnDrawCommand` to accept an immutable `BatchDrawCommandRef` instead, which it can use to read any relevant data from the draw command. Instead of being able to modify the draw command (which was bug prone to begin with, since modifying some parts like the flags would not actually work), it now returns the `BatchMaterialID` the draw should be performed with, or a null material ID in case the draw should be skipped. The mesh can no longer be modified, but this was not actually being used in practice, and would not be well defined for procedural draw types.

The code which needs to know the submesh index can now check if the draw actually has one, and default to zero in case it doesn't.