HDRP Composition Layer fixes
Fixes UUM-84610.

The main issue is here:

https://github.cds.internal.unity3d.com/unity/unity/blob/671a19ca55bb3d0eea4360949aa30d23ab69bf4a/Packages/com.unity.render-pipelines.high-definition/Runtime/Compositor/CompositionLayer.cs#L245-L250

Removing the `!enabled` condition fixes the issue. If the layer position on the stack is 0, it should always clear, regardless of the value of `enabled`. Otherwise, the first image might have garbage values on some platforms, which can cause visual issues (in this case on MacOS).

All the other changes are related to checking for null pointers, to avoid raising null pointer exceptions. I encountered several of these while fiddling around with the settings of the composition layer.