Use correct per-eye world space position when sampling APV in HDRP
https://jira.unity3d.com/browse/UUM-84382

This fix addresses a bug where the appearance of APV is noticeably different on each eye when using Single Pass Instanced XR rendering with Camera Relative Rendering enabled. It does not repro with Multi Pass XR rendering or with Camera Relative Rendering disabled.

The cause is that most of HDRP's lightloop uses "center eye" position for applying lights, i.e. the position halfway between each eye. It does so by modifying the input worldspace camera-relative fragment position to be the center eye position. The code that samples APV then erroneously passes this center eye position to `GetAbsolutePositionWS(..)` and uses the resulting position to tap the 3D texture. This is just wrong, `GetAbsolutePositionWS(..)` expects per-eye worldspace camera-relative position as input, but we give it center eye position, yielding bogus results.

Unlike the rest of the lightloop, APV doesn't need to use center eye position for sampling. It can use the actual per-eye worldspace (non-camera-relative) fragment position. My fix is to keep the unmodified camera-relative worldspace position around, pass it to `GetAbsolutePositionWS(..)` and use the result to sample APV.

Here's a before and after. The difference is subtle, but notice that the penumbra width is quite different on each before the fix.
![gifmaker_me (1)](https://media.github.cds.internal.unity3d.com/user/2392/files/eb43dc4f-f6ac-4efb-9347-e1befb0c3898)