UUM-86959: Fix environement lighting not updated when running in play…
JIRA: https://jira.unity3d.com/browse/UUM-86959
Slack: https://unity.slack.com/archives/C02FVPP6M34/p1731665800856559
Fix ambient probe not being updated when changing its intensity when running in play mode with static batching and GPU resident drawer enabled. Instances that are part of a static batch where explicitely discarded in the ProbesUpdateJob, which is incorrect since the ambient probe affects all gameobjects (static or not). The mistake probably happened because the job was written to be very similar to the TransformUpdateJob where it is valid to discard instances that are part of static batch