[Port] [2022.3] Sg/uum 76254 empty group paste location
Jira: https://jira.unity3d.com/browse/UUM-76254

Group position is adjusted as elements are added to it. However, if the group is empty, its position remains unchanged and is pasted at the original location. This has been fixed so that the group's position is adjusted when pasting, even if the group is empty.

While fixing this issue, I discovered and resolved a problem where the position of a newly created group was not being set initially


Before
![Before](https://media.github.cds.internal.unity3d.com/user/761/files/2a5877b8-75ae-4944-8290-f21ae3dfa098)

After
![After](https://media.github.cds.internal.unity3d.com/user/761/files/0dbdf1f9-f095-4c5f-bfb9-c23f406c7576)