Move Graphics Test Framework to Standard Path
This PR moves the Graphics Test Framework package to a path that includes the hyphen in the name. This is done in order to make the path standardized to the other existing paths we have. As discussed on Slack, renaming the package cannot happen right now, but moving it to a different path is okay.

https://jira.unity3d.com/browse/GSDET-396