The Baked Lightmap debug doesnt work with HDRP terrains
This PR fixes not to overwrite the texCoord1 in the lightmap debug mode.
For the lightmap debug mode, the texCoord1 is used for the lightmap UV. But it was overwritten to texCoord0 since terrain is always using uv0.

So, this PR fixed to use the texCoord1 value itself in the debug mode.

Before fixing:
![image](https://github.cds.internal.unity3d.com/unity/unity/assets/6282/13490a9e-2399-42b2-a235-f40828149e44)

After fixing:
![image](https://github.cds.internal.unity3d.com/unity/unity/assets/6282/0fa3776e-a244-45ef-8eaf-712a468cc37b)