[Backport 2022.3] Revert Changes from UUM-29958
I made a fix for https://jira.unity3d.com/browse/UUM-68986 that was too aggressive - clearing the material property block after each blit actually clears properties that are needed between blits, in the case for example of edge detection. This created another bug (https://jira.unity3d.com/browse/UUM-85970) where TexelSize property was cleared between two blits in the same pass.

Luckily, in the meantime, a better fix was implemented in Unity 6 and 7 (#51608). This PR backports it to 2022.3 (after talking to the original owner, there's should be no reason not to backport it)
This works better because it doesn't clear property blocks after each blit, only when a RenderTargetIdentifier is used, which is less aggressive than the fix I had in the first place