Shadergraph support for spacewarp
This PR's purpose is to add support for spacewarp, an XR feature.

Spacewarp works by drawing motions vectors and depth in specific textures, during a new specific pass (XRDepthMotionPass). So, this requires a new subshader pass to be added to shaders generated by shadergraph, if desired.

The PR adds the necessary code for a new checkbox, as well as code to add the new subshader pass to both Lit and Unlit targets.