DOCG-5984 Improved the Custom post-processing section, migrated the Fullscreen Master Stack pages.
Added a reference to the custom renderer feature page.
Added a reference to custom-post-processing.md on the customizing-urp.md page.
Implemented the same structure as in the 6000 version.
Copied Fullscreen Master Stack pages, the feature is implemented in 2022.3 and was not documented in URP.
Converted the formatting from the main manual to URP.