[VFX] Enable instancing support for VFX using GPU events
In VFX graph, GPU events allow particles to spawn particles into another system.
Due to the complexity of handling a list of variable events for different instances, it was excluded from the first instancing pass, disabling instancing for VFX assets that used GPU events.

<img width="860" alt="image" src="https://media.github.cds.internal.unity3d.com/user/2805/files/5c4049c2-03d6-433a-9c67-b958fae40e05">

This PR fixes that problem, allowing instancing regardless of the use of GPU events.
To do so, we need to build a prefix sum of the event count for each instance. Memory for this prefix sum has been reserved in the event count buffer. Currently it contains:
- Event counts: Number of events this frame. 1 per instance
- Total event count: Total number of events from the start. 1 per instance
- Prefix sum(NEW): Accumulated sum of events this frame, grouped by split group. 1 per instance
- Source index: Index where the source data for this event is located. 1 per event, per instance
[event counts (N) | total event counts (N) | prefix sum (N) | source index (N * max event count) ]

On the parent system:
- Each instance produces events and keeps track of its event count and event total count
- The batch update uses a compute shader to accumulate the event count for the entire batch
- After the parent system, FillIndirectArgs compute uses the accumulated event count to produce the dispatch indirect data for children Init

On the child system:
- Split in groups if needed (due to exposed textures, etc)
- Build a prefix sum for every group with the event counts of their instances
- Execute one Init for each split group with DispatchIndirect, with enough threads for the total event count
- Init will use the prefix sum to identify which thread uses which instance