[ShaderGraph] Prevent issues from multi-target property conflicts
Canvas default properties for stencil can conflict with HDRP. However; in this particular case, the property type is the same, so the normal deduplication that occurs is fine as a matter of coincidence. Suppressing the error because it doesn't have any practical impact for users nor is it possible for a user to correct.