Bicubic lightmap sampling for URP and HDRP
Jira: https://jira.unity3d.com/browse/GFXFEAT-813

This PR adds the ability to use bicubic sampling for lightmaps. Bicubic sampling smoothens high-frequency detail by doing more texture taps. This is especially helpful for low resolution lightmaps. Unlike bilinear filtering, which does 4 taps in a 2x2 shape (at the hardware level) and linearly interpolates them, bicubic takes 16 samples in a 4x4 shape and interpolates them using cubic splines. I've reused an existing implementation in SRP core. It uses a technique that does 4 hardware bilinear taps and uses them to get the same information as with 16 taps (see [this post](https://vec3.ca/bicubic-filtering-in-fewer-taps/) for more info).

The sampling strategy is implemented for both regular lightmaps, directional lightmaps and shadowmask textures. It has been exposed as a **global** setting, under the graphics tab. The main motivation for exposing it as a global setting is to not increase the shader variant count. A keyword, `LIGHTMAP_BICUBIC_SAMPLING` was added, but since the setting is global, we either only have variants that use the keyword, or only have variants that don't use the keyword. I've made sure to encode this logic into the variant stripping and prefiltering codepaths of each SRP. The setting is disabled by default.

![Unity_3DeLAXQrq2](https://media.github.cds.internal.unity3d.com/user/2392/files/29103423-166f-4b65-8415-4b9bdcc8b04e)