[Port] [6000.0] [HDRP] Adding cave scene in water samples
Following some discussions with users, we decided to add a small sample scene to use a bit more in depth the caustics and deformation texture.

https://github.cds.internal.unity3d.com/unity/unity/assets/1764/8ae9c5df-511b-4230-9b9f-a0dcdda7d113

Here's the 3 showcases : 
![image](https://github.cds.internal.unity3d.com/unity/unity/assets/1764/02e3bf9a-63bc-4358-a70b-96850eb61c15)

This basically adds very little to the water sample in terms of size since there's no new mesh or textures. 
Before : 23.47mb
After : 24.28mb