[Port] [6000.0] [SRP][Volume]Multiple fixes for being able to properly perform reset
Fix [UUM-77760](https://jira.unity3d.com/browse/UUM-77760) and additional issue around reset of default VolumeProfile:
- Fixed Reset attempt crash when resetting for pipeline not in use
- Fixed Reset nullifying when resetting for pipeline not in use instead of set back default values
- Fixed additional separators for HDRP version due to LookDev volume

To fix all of this, UITK conversion of the ImGUI VolumeProfile foldout field is done here.
Each SRPGraphicsSettings (URP, HDRP) have a Reset implementation to not nullify anymore.
On URP side, for the Reset to work, introduction of a new editor resource that load a default profile from package in Editor. The default vlume used is the one produced in the 3D template.