UUM-87651 : GPU Resident Drawer shadow caster cullingmask fix.
This PR fixes a discrepancy in URP between using GPU Resident Drawer when disabling shadow casting of certain objects for certain lights. In URP this can be done with a combination of game object layer on the shadow casting object and correct configuration of the light's Culling Mask to exclude that layer. Doing this enables correct lighting of that object from the light, without that object being considered as a shadow caster.

This combination wasn't working with GPU Resident Drawer because we were not taking into account the light's culling mask in the BatchRendererGroup's culling configuration before being sent to the OnPerformCulling, in the case of shadow caster culling.

As a side note, disabling shadow casting while retaining lighting on objects for specific lights can also be done with the Custom Shadow Layer functionality of both URP and HDRP. Consequently, this PR also modified the BatchRendererGroup_URP/MacroBatcher and BatchRendererGroup_HDRP/MacroBatcher graphics tests to check both techniques, using the first technique for URP and the Custom Shadow Layer technique for HDRP.