Land new Graphics Test Framework
This pull request introduces the first iteration in a major refresh of the Graphics Test Framework. The purpose of the changes is to create a modular, extensible base upon which to create a flexible test framework. 

The Graphics Test Framework has three core pseudo-modules: a graphics test builder, a runtime graphics test creator and the ImageAssert class. The first two have been refactored in this PR, with the third only receiving minor modifications.

Overall, the code has been moved around and breaking changes have been made. Several classes that existed before are now no more. **We have not yet updated any SRP test projects to use the new simplified workflow.**

Here is a summary of the changes. Please see the added / edited documentation for a full rundown of how the new setup works.


### Major Changes

1. Removed package dependencies which were not necessary / were slowing things down.
2. Now no longer able to run tests with the 'EditorBuildSettings' context. Need to explicitly tell the framework which scenes to look for.
3. Core APIs have been rewritten to be A LOT more concise.
4. Reference image loading and unloading is now possible (and doesn't happen until it's needed)
5. Asset bundle loading has been refactored and will continue to be so.

### SetupGraphicsTestCases becomes GraphicsTestBuilder

We have replaced the monolithic class `SetupGraphicsTestCases` and its even more monolithic `Setup` function with the `GraphicsTestBuilder` module and class.

The `GraphicsTestBuilder` is now an extensible class which can be modified using custom settings, custom build steps and/or entire custom build pipelines. Each of these allows the user an increasing level of control over the graphics test build process, from the location of reference image files, to custom build steps to be run as soon as we have a list of graphics test cases to build for.

Note that this is a first iteration of this functionality and it is yet to be fully integrated with some of the new features below. Regardless, it is a big step in modularizing the builder.

By default, the `GraphicsTestBuilder` class does not need to be called by the user. The `GraphicsTestBuildSettings` serializable object (which will be created and saved to the `Resources` folder upon first build if it does not exist) defines a boolean property `AutoBuildTestCases` which is true by default. This will cause test cases to be built upon running tests.

Of course, a user can disable this and control the entire build process and timing themselves.

### Scriptable Graphics Test Builder

We have added the abstract class `GraphicsPrebuildSetupAttribute`. Inheriting from this class and adding this attribute to a test fixture (or test) when it is connected to a `GraphicsTest` will cause the contained code (under the Setup method) to be run right before building Graphics Tests. Only one of each type of setup will be run as it's assumed it will run some global setup code.

### Graphics Test Case Collector

We have added a class which listens to events of `GraphicsTestCase` objects being created by the test framework. This allows us to access information about the test cases and the attached pre-build attributes. This feature is not being used to its full potential yet, but will allow us to fully abstract what a GraphicsTestCase is and allow us to decouple them from 'scenes' and the build settings in future PRs.

### CLI Test Build Filtering

We have enabled a bare-bones test build filtering through the 'testFilter' command-line flag. Until now it was not possible to control which tests are build (and thus limit build times) but now it is! You're welcome. There is no support for wild-cards yet, but it is in the works to do so.

---

### UseGraphicsTestCasesAttribute and CodeBasedGraphicsTestAttribute become GraphicsTestAttribute

The `[GraphicsTest]` attribute, a much shorter version of the existing attributes, combines the powers of its two predecessors in one. This attribute is the only markup that one needs to run GraphicsTests. 

By default, `[GraphicsTest]` will create a test with the 'DefaultGraphicsTestCaseSource` which will assume that you're just writing a normal unit test and just need access to a 'GraphicsTestCase' parameter (for that automatic reference image management perhaps?)

Otherwise, a user may extend the GraphicsTestCaseSource class to create a custom provider / builder of GraphicsTestCase objects (which will then happily be built into your tests as the first parameter!) We have included the classic `SceneGraphicsTestCaseSource` and its own attribute `[SceneGraphicsTest]` as it's the most common use case at the moment, but rest assured that you can write your own if you'd like to and reference them in your tests!

`GraphicsTest` and its descendants are also compatible with parametric tests (using the `ValueSource` attribute or the `GraphicsTestCaseAttribute` or `GraphicsTestCaseSourceAttribute`) and compatible with both `Test` and `UnityTest`. This will allow for much faster, flexible less duplicated image comparison tests.

---

### Reference Image

We have introduced a new class called `ReferenceImage` to encapsulate the concept of a `ReferenceImage` which has up to this point been just bare Texture2Ds. This feature is already being used in the framework but due to the scoping down of this PR to not include API changes to ImageAssert, it has not completely replaced the current implementation.


---

### IgnoreGraphicsTest Attribute

The attribute is meant to be used in a similar way as the [IgnoreAttribute], but with the added utility of being able to ignore specific patterns of GraphicsTestCases.

The first argument is a regular expression. The second argument is the reason for ignoring the test. The following (optional) arguments are arrays of the enums that make up GraphicsTestPlatform. Due to the limitations of attributes, this was the only way to implement this cleanly and I'm pretty satisfied.

For example, to ignore test 1234-Something-something on Linux Vulkan you'd write:

[IgnoreGraphicsTest("1234", "Reason", new RuntimePlatform[] { RuntimePlatform.LinuxEditor, RuntimePlatform.LinuxStandalone }, new GraphicsDeviceType[] { GraphicsDeviceType.Vulkan })]

Any omitted values are assumed to be any value. Therefore the values provided define the superset of platforms that the ignore matches. The platform arguments can be entirely omitted to permit matching any platform (thereby ignoring the test universally).

The attribute can only work at the test level at the moment, not at the test fixture level as the documentation may say.