[UUM-84459] Add a condition to the Receiver Motion Rejection function to ensure that the pixel has actually moved.
Jira : [UUM-84459](https://jira.unity3d.com/browse/UUM-84459)

Problem:
When the Screen Space Global Illumination is in Ray Tracing mode, there is an issue where the "Receiver Motion Rejection" option is enabled. It discards history when skinned meshes have not actually moved.

Cause:
In the ValidateHistory kernel of TemporalFilter.compute, the HISTORYREJECTIONFLAGS_MOTION flag is determined solely based on the stencilValue. For skinned meshes, the stencilValue is always updated in the ObjectsMotionVector pass, which causes the HISTORYREJECTIONFLAGS_MOTION flag to be set even if the pixels have not moved.

Solution:
Instead of checking only the stencilValue, modify the process to check the motion vector (velocity) of the pixel and set the flag only when this value is not float2(0, 0).