[URP2D] Remove redundant Shadow2DUnshadowGeometry Blend [PLATGRAPH-4108]
URP 2D shader `Shadow2DUnshadowGeometry` enables Blend and specifies `ZERO` for Destination Blend Factor, and `SrcColor` for Source Blend Factor. Also, the output of this shader is half4(1,1,1,1).

This means that the output of this blend operation is:

     srcColor * srcFactor + dstColor * dstFactor = (1,1,1,1) . (1,1,1,1) + ZERO = (1,1,1,1)

...which means the Blend operation is not needed.

On some platforms for some render target formats, blend throughput is slower than "fill throughput" (rendering the pixels); so to improve performance it is better to avoid Blend when that is possible like in this case.

Jira ticket: [PLATGRAPH-4108](https://jira.unity3d.com/browse/PLATGRAPH-4108)