[Port] [6000.0] [SRPF] Destroy Renderers when they created in the Build process without RP creation
This issue appeared after https://github.cds.internal.unity3d.com/unity/unity/pull/54848 PR landed. It prevented RP creation and rendering from scene loading in the middle of the Building process. 

Now whenever we build we may have a situation when there's no current pipeline as no one called rendering. It should be okay but some of the API called GetRenderer and didn't dispose them after. To prevent this we check if URP is not ready yet and Destroy all existing Renderers.